update phân quyên

+++ app/Http/Controllers/BackEnd/Contacts/ContactController.php
@@ -1,10 +1,12 @@
 <?php
 
 namespace App\Http\Controllers\Backend\Contacts;
+use App\Constant\PermissionConstant;
 use App\Models\Contact;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Lang;
+use Illuminate\Support\Facades\Auth;
 class ContactController extends Controller
 {
     protected $contacts;
@@ -16,8 +18,10 @@ class ContactController extends Controller
 
     public function index()
     {
+        if(Auth::user()->can(PermissionConstant::CONTATCS_PERMISSION_LIST)) {
         $contacts = $this->contacts->sortable()->paginate(3);
         return view('backend.admin.contacts.index', compact('contacts'));
+        }
     }
 
   
@@ -25,15 +29,19 @@ class ContactController extends Controller
  
     public function view(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::CONTACTS_PERMISSION_VIEW)) {
         $contacts = $this->contacts->find($request->id);
         return view('backend/admin/contacts/view', compact('contacts'));
+        }
     }
 
     public function delete(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::CONTACTS_PERMISSION_DELETE)) {
         $contacts = $this->contacts->find($request->id);
         $contacts->delete();
         return redirect()->route('backend.admin.contacts.show')->with('success', Lang::get('message.delete', ['model' => 'Ý kiến khách hàng']));
+        }
     }
 
     public function contactsDelete(Request $request)
